# MH-13277 improve scheduler performance