--- a/etc/org.opencastproject.scheduler.impl.SchedulerServiceImpl.cfg
+++ b/etc/org.opencastproject.scheduler.impl.SchedulerServiceImpl.cfg
@@ -4,7 +4,7 @@
 # Default: 60 seconds
 #last_modified_cache_expire=60
 
-# Time in seconds to wait before cleaning up old transactions. Note that this will terminate running transactions for
-# the sake of releasing unfinished but abandoned transactions to avoid locking schedules forever.
-# Default: 864000 seconds (10 days)
-#transaction_cleanup_offset=864000
+# Whether to enable maintenance mode or not. Maintenance mode will deactivate the scheduler. This is only needed when
+# migrating data of the scheduler.
+# Default: false
+#maintenance=false