--- a/modules/external-api/src/main/java/org/opencastproject/external/endpoint/EventsEndpoint.java
+++ b/modules/external-api/src/main/java/org/opencastproject/external/endpoint/EventsEndpoint.java
@@ -1763,8 +1763,7 @@ public class EventsEndpoint implements ManagedService {
           Opt.none(),
           Opt.none(),
           caConfig,
-          Opt.none(),
-          SchedulerService.ORIGIN);
+          Opt.none());
     } catch (SchedulerConflictException e) {
       final List<MediaPackage> conflictingEvents = getConflictingEvents(
           schedulingInfo.merge(technicalMetadata), agentStateService, schedulerService);