--- a/modules/index-service/src/main/java/org/opencastproject/index/service/impl/IndexServiceImpl.java
+++ b/modules/index-service/src/main/java/org/opencastproject/index/service/impl/IndexServiceImpl.java
@@ -957,8 +957,7 @@ public class IndexServiceImpl implements IndexService {
         eventHttpServletRequest.setMediaPackage(mediaPackage);
         try {
           schedulerService.addEvent(start.toDate(), start.plus(duration).toDate(), captureAgentId, presenterUsernames,
-                  mediaPackage, configuration, (Map) caProperties, Opt.<Boolean> none(), Opt.<String> none(),
-                  SchedulerService.ORIGIN);
+                  mediaPackage, configuration, (Map) caProperties, Opt.<Boolean> none(), Opt.<String> none());
         } finally {
           for (MediaPackageElement mediaPackageElement : mediaPackage.getElements()) {
             try {
@@ -973,7 +972,7 @@ public class IndexServiceImpl implements IndexService {
         List<Period> periods = schedulerService.calculatePeriods(rRule, start.toDate(), end.toDate(), duration, tz);
         Map<String, Period> scheduled = new LinkedHashMap<>();
          scheduled = schedulerService.addMultipleEvents(rRule, start.toDate(), end.toDate(), duration, tz, captureAgentId,
-                presenterUsernames, eventHttpServletRequest.getMediaPackage().get(), configuration, (Map) caProperties, Opt.none(), Opt.none(), SchedulerService.ORIGIN);
+                presenterUsernames, eventHttpServletRequest.getMediaPackage().get(), configuration, (Map) caProperties, Opt.none(), Opt.none());
         return StringUtils.join(scheduled.keySet(), ",");
       default:
         logger.warn("Unknown source type {}", type);
@@ -1220,7 +1219,7 @@ public class IndexServiceImpl implements IndexService {
         try {
           schedulerService.updateEvent(event.getIdentifier(), Opt.<Date> none(), Opt.<Date> none(), Opt.<String> none(),
                   Opt.<Set<String>> none(), Opt.some(mediaPackage), Opt.<Map<String, String>> none(),
-                  Opt.<Map<String, String>> none(), Opt.<Opt<Boolean>> none(), SchedulerService.ORIGIN);
+                  Opt.<Map<String, String>> none(), Opt.<Opt<Boolean>> none());
         } catch (SchedulerException e) {
           logger.error("Unable to remove catalog with flavor {} by updating scheduled event {} because {}",
                   flavor, event.getIdentifier(), getStackTrace(e));
@@ -1295,7 +1294,7 @@ public class IndexServiceImpl implements IndexService {
         try {
           schedulerService.updateEvent(id, Opt.<Date> none(), Opt.<Date> none(), Opt.<String> none(), presenters,
                   Opt.some(mediaPackage), Opt.<Map<String, String>> none(), Opt.<Map<String, String>> none(),
-                  Opt.<Opt<Boolean>> none(), SchedulerService.ORIGIN);
+                  Opt.<Opt<Boolean>> none());
         } catch (SchedulerException e) {
           logger.error("Unable to update scheduled event {} with metadata {} because {}",
                   id, RestUtils.getJsonStringSilent(metadataList.toJSON()), getStackTrace(e));
@@ -1361,7 +1360,7 @@ public class IndexServiceImpl implements IndexService {
         try {
           mediaPackage = authorizationService.setAcl(mediaPackage, AclScope.Episode, acl).getA();
           schedulerService.updateEvent(id, Opt.none(), Opt.none(), Opt.none(), Opt.none(), Opt.some(mediaPackage),
-                  Opt.none(), Opt.none(), Opt.none(), SchedulerService.ORIGIN);
+                  Opt.none(), Opt.none(), Opt.none());
         } catch (SchedulerException | MediaPackageException e) {
           throw new IndexServiceException("Unable to update the acl for the scheduled event", e);
         }
@@ -1992,7 +1991,7 @@ public class IndexServiceImpl implements IndexService {
           case SCHEDULE:
             logger.info("Update scheduled mediapacakge {} with updated comments catalog.", event.getIdentifier());
             schedulerService.updateEvent(event.getIdentifier(), Opt.none(), Opt.none(), Opt.none(), Opt.none(),
-                    Opt.some(mediaPackage), Opt.none(), Opt.none(), Opt.none(), SchedulerService.ORIGIN);
+                    Opt.some(mediaPackage), Opt.none(), Opt.none(), Opt.none());
             break;
           default:
             logger.error("Unkown event source {}!", event.getSource());
@@ -2054,7 +2053,7 @@ public class IndexServiceImpl implements IndexService {
 
     schedulerService.updateEvent(eventId, Opt.<Date> none(), Opt.<Date> none(), Opt.<String> none(),
             Opt.<Set<String>> none(), Opt.<MediaPackage> none(), Opt.<Map<String, String>> none(),
-            Opt.<Map<String, String>> none(), Opt.some(Opt.some(optout)), SchedulerService.ORIGIN);
+            Opt.<Map<String, String>> none(), Opt.some(Opt.some(optout)));
     logger.debug("Setting event {} to opt out status of {}", eventId, optout);
   }