… docs/guides/admin/docs/releasenotes.md
… docs/guides/admin/docs/upgrade.md
… docs/guides/developer/docs/scheduler.md
… modules/admin-ui/src/test/java/org/opencastproject/adminui/endpoint/TestEventEndpoint.java
… modules/external-api/src/test/java/org/opencastproject/external/endpoint/TestEventsEndpoint.java
… modules/index-service/src/test/java/org/opencastproject/index/service/impl/IndexServiceImplTest.java
… modules/index-service/src/test/java/org/opencastproject/index/service/message/SchedulerMessageReceiverImplTest.java
… modules/message-broker-api/src/main/java/org/opencastproject/message/broker/api/scheduler/SchedulerItem.java
… modules/message-broker-api/src/main/java/org/opencastproject/message/broker/api/scheduler/SchedulerItemList.java
… modules/migration/pom.xml
… modules/migration/src/main/java/org/opencastproject/migration/EventIdPK.java
… modules/migration/src/main/java/org/opencastproject/migration/SchedulerMigrationService.java
… modules/migration/src/main/resources/META-INF/persistence.xml
… modules/migration/src/main/resources/OSGI-INF/scheduler-migration-service.xml
… modules/migration/src/test/java/org/opencastproject/migration/SchedulerMigrationServiceTest.java
… modules/migration/src/test/resources/log4j.properties
… modules/scheduler-api/src/main/java/org/opencastproject/scheduler/api/SchedulerService.java
… modules/scheduler-api/src/main/java/org/opencastproject/scheduler/api/SchedulerTransactionLockException.java
… modules/scheduler-impl/pom.xml
… modules/scheduler-impl/src/main/java/org/opencastproject/scheduler/endpoint/SchedulerRestService.java
… modules/scheduler-impl/src/main/java/org/opencastproject/scheduler/impl/CaptureNowProlongingService.java
… modules/scheduler-impl/src/main/java/org/opencastproject/scheduler/impl/RuntimeNotFoundException.java
… modules/scheduler-impl/src/main/java/org/opencastproject/scheduler/impl/SchedulerServiceDatabase.java
… modules/scheduler-impl/src/main/java/org/opencastproject/scheduler/impl/SchedulerServiceImpl.java
… modules/scheduler-impl/src/main/java/org/opencastproject/scheduler/impl/SchedulerUtil.java
… modules/scheduler-impl/src/main/java/org/opencastproject/scheduler/impl/TransactionCleaner.java
… modules/scheduler-impl/src/main/java/org/opencastproject/scheduler/impl/persistence/EventIdPK.java
… modules/scheduler-impl/src/main/java/org/opencastproject/scheduler/impl/persistence/ExtendedEventDto.java
… modules/scheduler-impl/src/main/java/org/opencastproject/scheduler/impl/persistence/SchedulerServiceDatabaseImpl.java
… modules/scheduler-impl/src/main/java/org/opencastproject/scheduler/impl/persistence/TransactionDto.java