This patchset improves the scheduler performance as part of the opencast crowdfunding 2018. As a benchmark, we scheduled 50k events and then added 6 events on top. Previously, adding the 6 events took between 19 and 20 seconds. Of these 20 seconds, the conflict check took about 80% of the time. After our changes, it takes about 2 seconds to add the 6 events on top. Of these 2 seconds, the conflict check now only takes a very small percentage of time. The benchmark was done using Opencast with MariaDB on an AMD FX(tm)-8350 Eight-Core Processor with 32GB RAM and an SSD.

The following changes were done to achieve the goal of faster conflict checks and higher throughput when inserting events:

Remove Unused YAGNI Code
-----------------------------------------

- The "transactional" feature of the scheduler was only used for the migration from opencast 3.x to 4.x and introduced a lot of complexity. In order to improve the maintainability of opencast, we decided to remove this code. If anyone wants to utilize it someday, it can be restored from the git history.
- The "blacklist" feature of the scheduler wasn't fully implemented. We removed it as well, again for the sake of maintainability.

Store Properties of Scheduled Events in a Dedicated Table
--------------------------------------------------------------------------------

Properties such as start date, end date, capture device id, of scheduled events were stored as asset manager properties which are basically key-value pairs which cannot be stored efficiently using a relational DBMS. For this reason, the conflict check was done in-memory, which was a workaround for the slow database queries. However, this still was slow since all events had to be loaded into memory before performing the conflict check. The scheduler now uses a proper "SQL-friendly" table to store these properties. This drastically improves the conflict check performance (which can be done on the database again) as well as the insertion speed when adding new events.

Reduce Number of ActiveMQ Messages When Adding Events
-------------------------------------------------------------------------------------

When scheduling multiple events, one message per event was sent over ActiveMQ. This seemed to cause some buffers filling up and blocking the message sender which in turn slowed down event insertion. We changed the behavior to send just one message for a whole bunch of events, which solves this problem.

Parallelize Adding Multiple Events
-----------------------------------------------

When scheduling multiple events (using an RRULE), these events cannot have conflicts with each other. This allows us to parallelize event insertion in this case, which also improves the throughput.

Other Changes
---------------------

Additionally, we improved the test coverage of the scheduler and added a migration which is necessary because of the database schema changes,