@JamesUoM 2018-12-07
Is there any way to store custom properties, in scheduled events? We had planned to tighter integrate our Participation Manager (opt-out) with the using Asset managers properties to store addition choices like editing, camera and or screen recording, without having to parse the workflow and capture agent configs.