@krinnewitz 2018-12-07
It depends.If you know the keys and types of your properties in advance, you can just add  one column per property to oc_scheduled_extended_event and store them there.

If the properties you want to store are arbitrary and don't have to be queried individually, you can add another string column, serialize the properties using gson and store them in this column. For reading those properties, you just read the string from the database and de-serialize them again. Just like we did it for the capture agent and workflow properties. 

The problem with the workflow and capture agent properties was, that they can consist of arbitrary key-value pairs, so we couldn't create one column per property. But since these properties werent't queried individually, we could just serialize them and store them as a whole.