@staubesv 2018-12-07
@JamesUoM & @krinnewitz Opencast 6.x introduces a new processing settings persistent layer that uses the asset manager to store workflow properties. At some point in the future, it likely makes sense to remove the workflow properties from the scheduler table and make the scheduler use the processing settings persistent layer of the asset manager. At the end of the day, a scheduled event is a event and it would be perfectly fine to store the workflow properties just for the event (vs. scheduled event).