@staubesv 2018-12-07
@krinnewitz Please also adopt the page https://docs.opencast.org/develop/developer/scheduler/