@krinnewitz 2018-12-08
Thanks @staubesv. I updated the documentation. 

Regarding the workflow properties: Maybe we could change that in a follow-up PR if there are not too many review findings with this one. However, the problem with storing anything in the asset manager properties is, that even just insertion is quite slow. So when scheduling multiple events at once, that could already be a problem.