@gregorydlogan 2019-01-09
I've tagged this as a feature, even though it's not *really* a feature.  It *does* however warrant the extra testing that comes with being a feature IMO.