@lkiesow 2019-01-13
Sorry for the late start of the review.

> When scheduling multiple events (using an RRULE), these events cannot have conflicts with each 
> other. This allows us to parallelize event insertion in this case, which also improves the throughput.

Does Opencast's internal processing of RRULES differ from what's possible in iCalendar? Because I think this would be a completely valid schedule with overlapping events using an RRULE (12 events á 2h scheduled hourly):

```ical
BEGIN:VEVENT
DTSTAMP:20190113T185826Z
UID:20190113T185826Z-1469192696
DTSTART;TZID=Europe/Berlin:20190201T120000
DTEND;TZID=Europe/Berlin:20190201T140000
RRULE:FREQ=HOURLY;INTERVAL=1;COUNT=12
SUMMARY:Test
DESCRIPTION:Some overlapping test events
END:VEVENT
```