@pmiddend 2019-01-14
@lkiesow You're right. I hope the admin UI actually prevents this from happening, but surely there are other ways to "inject" such an RRULE into the system. Either way, I've added some code to check those overlaps.