@lkiesow 2019-01-27
I've tested this in a production-like environment and automated scheduling of ~400 events worked very well. The only problem I really noticed was the weird naming I already found in the code earlier when manually scheduling multiple events. Apart from that, everything seems to work.