@krinnewitz 2019-01-28
> Hi, since this pull has a lot of them, can someone explain why it is good for Opencast to make new iterations of NotFoundException, etc, that do not extend plain old Java NotFoundExceptions, etc? I do not understand the gained benefits over the complexity added.

Hi @karendolan I think in most cases it makes sense to not throw exceptions which are part of the java API since you want to handle exceptions thrown by the java class library differently from the ones thrown by your own code. And in most cases you want to express a different situation than the "standard" java exception stands for.

For example here is a snippet from the javadoc of `javax.ws.rs.NotFoundException`:

> A runtime exception indicating a resource requested by a client was not found on the server.

That may not make sense in all places.